+++ docs/source/changelog.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+2021.08.1
+---------
+
+Enhancements
+
+    HTTTP
+
+    - get_file/put_file APIs now support callbacks
+    - New put_file API for transferring data to the remote server (chunked)
+    - Customizable client initializers (through passing ``get_client`` argument)
+    - Support for various checksum / fingerprint headers in ``info()`` (``ETag``, ``MD5-Checksum``, ``Digest``)
+
 2021.07.0
 ---------